requirements.txt
Django==2.2.7
psycopg2==2.8.4
psycopg2-binary==2.8.4
django-import-export==1.2.0
django-ckeditor==5.8.0